requirements.txt
Pillow==9.3.0
Pillow==10.0.0
torch==2.0.1+cu117
torchvision==0.15.2+cu117
